+++ frontend/src/components/assessment-management/editor/BasicInformation.tsx
