+++ frontend/src/components/assessment-management/editor/QuestionSummary.tsx
